Produce coverage info
